CustomFieldGraph: refactor (WIP)
