+++ front/app/modules/commercial/user_custom_fields/admin/components/CustomFieldGraphs/CustomFieldGraph.tsx
@@ -1,6 +1,5 @@
-// libraries
 import React from 'react';
-import { isEmpty, map, orderBy } from 'lodash-es';
+import { isEmpty } from 'lodash-es';
 import { Subscription, combineLatest } from 'rxjs';
 
 // intl
@@ -26,12 +25,11 @@ import BarChart, { DEFAULT_MARGIN } from 'components/admin/Graphs/BarChart';
 import { Box, colors } from '@citizenlab/cl2-component-library';
 
 // typings
-import { IUserCustomFieldData } from '../../services/userCustomFields';
+import { IUserCustomFieldData } from '../../../services/userCustomFields';
 import { IStream } from 'utils/streams';
 
 // services
 import {
-  IUsersByRegistrationField,
   usersByRegFieldStream,
   usersByRegFieldXlsxEndpoint,
   usersByGenderStream,
@@ -40,37 +38,13 @@ import {
   usersByBirthyearXlsxEndpoint,
   usersByDomicileStream,
   usersByDomicileXlsxEndpoint,
-  IUsersByGender,
-  IUsersByDomicile,
-  IUsersByBirthyear,
 } from 'modules/commercial/user_custom_fields/services/stats';
 
 // utils
 import { isNilOrError } from 'utils/helperUtils';
-import {
-  binBirthyear,
-  rename,
-  join,
-  convertDomicileData,
-  Series,
-} from '../../utils/data';
-import { fallbackMessages } from './AreaChart';
-
-// hooks
-import useUserCustomFields from '../../hooks/useUserCustomFields';
-
-type ISupportedDataType =
-  | IUsersByRegistrationField
-  | IUsersByGender
-  | IUsersByDomicile
-  | IUsersByBirthyear;
-
-const genderColors = {
-  male: '#5D99C6 ',
-  female: '#C37281 ',
-  unspecified: '#B0CDC4 ',
-  _blank: '#C0C2CE',
-};
+import createConvertAndMergeSeries, {
+  TConvertAndMergeSeries,
+} from './convertAndMergeSeries';
 
 const customFieldEndpoints = {
   gender: {
@@ -87,18 +61,13 @@ const customFieldEndpoints = {
   },
 };
 
-const joinTotalAndParticipants = (total: Series, participants: Series) =>
-  join(
-    rename(total, { value: 'total' }),
-    rename(participants, { value: 'participants' }),
-    { by: 'name' }
-  );
-
-interface Props {
+interface InputProps {
   customField: IUserCustomFieldData;
   currentProject: string | undefined;
   className?: string;
 }
+
+type Props = InputProps & InjectedIntlProps & InjectedLocalized;
 
 interface State {
   serie: any;
@@ -132,12 +101,10 @@ const CustomTooltip = ({
   return null;
 };
 
-export class CustomFieldsComparison extends React.PureComponent<
-  Props & InjectedIntlProps & InjectedLocalized,
-  State
-> {
+export class CustomFieldsGraph extends React.PureComponent<Props, State> {
   combined$: Subscription;
   currentChart: React.RefObject<any>;
+  convertAndMergeSeries: TConvertAndMergeSeries;
 
   constructor(props: Props) {
     super(props as any);
@@ -146,6 +113,15 @@ export class CustomFieldsComparison extends React.PureComponent<
     };
 
     this.currentChart = React.createRef();
+
+    const {
+      localize,
+      intl: { formatMessage },
+    } = props;
+    this.convertAndMergeSeries = createConvertAndMergeSeries({
+      localize,
+      formatMessage,
+    });
   }
 
   componentDidMount() {
@@ -163,71 +139,6 @@ export class CustomFieldsComparison extends React.PureComponent<
   componentWillUnmount() {
     this.combined$.unsubscribe();
   }
-
-  convertAndMergeSeries = (
-    totalSerie: ISupportedDataType,
-    participantSerie: ISupportedDataType
-  ) => {
-    const {
-      customField,
-      intl: { formatMessage },
-      localize,
-    } = this.props;
-
-    if (customField.attributes.code === 'birthyear') {
-      const options = { missing: formatMessage(messages._blank) };
-
-      const binnedTotal = binBirthyear(totalSerie.series.users, options);
-      const binnedParticipants = binBirthyear(
-        participantSerie.series.users,
-        options
-      );
-
-      return joinTotalAndParticipants(binnedTotal, binnedParticipants);
-    } else if (customField.attributes.code === 'gender') {
-      const res = Object.keys(genderColors).map((gender) => ({
-        total: totalSerie.series.users[gender] || 0,
-        participants: participantSerie.series.users[gender] || 0,
-        name: formatMessage(messages[gender]),
-        code: gender,
-      }));
-      return res.length > 0 ? res : null;
-    } else if (customField.attributes.code === 'domicile') {
-      const parseName = (key, value) =>
-        key in fallbackMessages
-          ? formatMessage(fallbackMessages[key])
-          : localize(value.title_multiloc);
-
-      const areas = (totalSerie as IUsersByDomicile).areas;
-      const resTotal = convertDomicileData(
-        areas,
-        totalSerie.series.users,
-        parseName
-      );
-      const resParticipants = convertDomicileData(
-        areas,
-        participantSerie.series.users,
-        parseName
-      );
-      const res = joinTotalAndParticipants(resTotal, resParticipants);
-
-      const sortedByParticipants = orderBy(res, 'participants', 'desc');
-      return sortedByParticipants;
-    } else {
-      const res = map(
-        (totalSerie as IUsersByRegistrationField).options,
-        (value, key) => ({
-          total: totalSerie.series.users[key] || 0,
-          participants: participantSerie.series.users[key] || 0,
-          name: localize(value.title_multiloc),
-          code: key,
-        })
-      );
-
-      const sortedByParticipants = orderBy(res, 'participants', 'desc');
-      return sortedByParticipants;
-    }
-  };
 
   resubscribe() {
     const { customField, currentProject } = this.props;
@@ -249,10 +160,14 @@ export class CustomFieldsComparison extends React.PureComponent<
       participantsStream.observable,
     ]).subscribe(([totalSerie, participantSerie]) => {
       if (!isNilOrError(totalSerie) && !isNilOrError(participantSerie)) {
+        const { code } = customField.attributes;
+
         const convertedAndMergedSeries = this.convertAndMergeSeries(
-          totalSerie as ISupportedDataType,
-          participantSerie as ISupportedDataType
+          totalSerie,
+          participantSerie,
+          code
         );
+
         this.setState({ serie: convertedAndMergedSeries });
       }
     });
@@ -328,42 +243,8 @@ export class CustomFieldsComparison extends React.PureComponent<
   }
 }
 
-const WrappedCustomFieldComparison = injectLocalize<Props>(
-  injectIntl<Props & InjectedLocalized>(
-    withTheme(CustomFieldsComparison as any) as any
+export default injectLocalize<InputProps>(
+  injectIntl<InputProps & InjectedLocalized>(
+    withTheme(CustomFieldsGraph as any) as any
   )
 );
-
-const CustomFieldGraphs = ({
-  participationMethods,
-  startAt,
-  endAt,
-  project,
-}) => {
-  const userCustomFields = useUserCustomFields({
-    inputTypes: ['select', 'multiselect', 'checkbox', 'number'],
-  });
-
-  return (
-    participationMethods !== ['information'] &&
-    startAt &&
-    endAt &&
-    !isNilOrError(userCustomFields) &&
-    userCustomFields.map(
-      (customField) =>
-        // only show enabled fields, only supported number field is birthyear.
-        customField.attributes.enabled &&
-        (customField.attributes.input_type === 'number'
-          ? customField.attributes.code === 'birthyear'
-          : true) && (
-          <WrappedCustomFieldComparison
-            customField={customField}
-            currentProject={project.id}
-            key={customField.id}
-          />
-        )
-    )
-  );
-};
-
-export default CustomFieldGraphs;
